+++ internal/app/cli/scan.go
@@ -30,7 +30,7 @@ import (
 func NewScanCommand(parent string, stderr *os.File, stdout *os.File, stdin *os.File) *cobra.Command {
 	var (
 		passthrough bool
-		only        []string
+		include     []string
 		aliases     map[string]string
 	)
 
@@ -40,14 +40,14 @@ func NewScanCommand(parent string, stderr *os.File, stdout *os.File, stdin *os.F
 		Example: "  lino pull database --table client | " + parent + " scan clients",
 		Args:    cobra.ExactArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			if err := scan(cmd, args[0], passthrough, only, aliases); err != nil {
+			if err := scan(cmd, args[0], passthrough, include, aliases); err != nil {
 				log.Fatal().Err(err).Int("return", 1).Msg("end SILO")
 			}
 		},
 	}
 
 	cmd.Flags().BoolVarP(&passthrough, "passthrough", "p", false, "pass stdin to stdout")
-	cmd.Flags().StringSliceVarP(&only, "only", "o", []string{}, "only scan these columns, exclude all others")
+	cmd.Flags().StringSliceVarP(&include, "include", "i", []string{}, "include only these columns, exclude all others")
 	cmd.Flags().StringToStringVarP(&aliases, "alias", "a", map[string]string{}, "use given aliases for each columns")
 
 	cmd.SetOut(stdout)
@@ -57,7 +57,7 @@ func NewScanCommand(parent string, stderr *os.File, stdout *os.File, stdin *os.F
 	return cmd
 }
 
-func scan(cmd *cobra.Command, path string, passthrough bool, only []string, aliases map[string]string) error {
+func scan(cmd *cobra.Command, path string, passthrough bool, include []string, aliases map[string]string) error {
 	backend, err := infra.NewBackend(path)
 	if err != nil {
 		return fmt.Errorf("%w", err)
@@ -65,7 +65,7 @@ func scan(cmd *cobra.Command, path string, passthrough bool, only []string, alia
 
 	defer backend.Close()
 
-	driver := silo.NewDriver(backend, nil, silo.WithKeys(only), silo.WithAliases(aliases))
+	driver := silo.NewDriver(backend, nil, silo.WithKeys(include), silo.WithAliases(aliases))
 
 	var reader silo.DataRowReader